Proper file subdivision
* Atlas is probably completed
* Explorer is not
